move `set_collection_id` to nft-transfer tests
