+++ pallets/ajuna-awesome-avatars/src/tests.rs
@@ -3563,25 +3563,6 @@ mod nft_transfer {
 	use sp_runtime::bounded_vec;
 
 	#[test]
-	fn set_collection_id_works() {
-		ExtBuilder::default().organizer(CHARLIE).build().execute_with(|| {
-			let collection_id = 369;
-			assert_ok!(AAvatars::set_collection_id(RuntimeOrigin::signed(CHARLIE), collection_id));
-			assert_eq!(CollectionId::<Test>::get(), Some(collection_id));
-		});
-	}
-
-	#[test]
-	fn set_collection_id_rejects_non_organizer_calls() {
-		ExtBuilder::default().organizer(ALICE).build().execute_with(|| {
-			assert_noop!(
-				AAvatars::set_collection_id(RuntimeOrigin::signed(BOB), 333),
-				DispatchError::BadOrigin
-			);
-		});
-	}
-
-	#[test]
 	fn can_lock_avatar_successfully() {
 		let season = Season::default()
 			.mint_logic(LogicGeneration::Second)
